rolled back moving of libraries
