+++ FunctionsSandboxLibrary/Log.js
