+++ FunctionsSandboxLibrary/encryptSecrets.js
