+++ hardhat.config.js
@@ -1,4 +1,4 @@
-require("./libraries/env-enc").config()
+require("./env-enc").config()
 require("@nomicfoundation/hardhat-toolbox")
 require("hardhat-contract-sizer")
 require("@openzeppelin/hardhat-upgrades")
