+++ package.json
@@ -18,7 +18,7 @@
     "format:fix": "prettier --write ."
   },
   "bin": {
-    "env-enc": "./libraries/env-enc/cli.js"
+    "env-enc": "./env-enc/cli.js"
   },
   "devDependencies": {
     "@chainlink/contracts": "0.5.1",
